e2e folder migrated to ts
